add test and move util to correct file
